Overview

Similar to pynx and others?, BcdiTrad implements projection-based algorithms in operator-style format. This means that the multiplication (*) and power (^) operators are used to apply operators to some current state. This may look like the following:

er = BcdiTrad.ER()
hio = BcdiTrad.HIO(0.9)
state = BcdiTrad.State(intensities, trues(size(intensities)))

(er * hio^20)^5 * state

This short script applies 20 HIO iterations and one ER iterations for a total of 5 times. This makes it easy to implement very complex recipes for phase retrieval algorithms.

API

Main.DummyDocs.State — Method

State(intensities, recSupport)

Create a reconstruction object. The intensities and a mask over reciprocal space indicating trusted intensities need to be passed in.

Main.DummyDocs.ER — Method

ER()

Create an object that applies an iteration of ER

Main.DummyDocs.HIO — Method

HIO(beta)

Create an object that applies an iteration of HIO

Main.DummyDocs.Shrink — Method

Shrink(threshold, sigma, state)

Create an object that applies shrinkwrap

Main.DummyDocs.Center — Method

Center(state)

Create an object that centers the current state